Switching to a .ufoz-based workflow
The ufo is no longer included in the repository, but instead the ufoz. So there's now an extract script (and a "compress.sh" to create the .ufoz before uploading) that converts before building.

I've also updated the pre-processing file to add NID entries for localized NID1, and Latin NID16 to keep things similar across languages.